Laboratorios/P1-2/P1/src/AlgoritmosTests.java
import static org.junit.jupiter.api.Assertions.*;

import org.junit.jupiter.api.Test;

class AlgoritmosTests {

	@Test
	void testRecursivePow1() {
		try {
			assertEquals(1, Algoritmos.recursivePow(5, 0));
		} catch (AssertionError e) {
			System.out.println("El resultado debería ser 1");
			throw e;
		}
	}

	@Test
	void testRecursivePow2() {
		try {
			assertEquals(125, Algoritmos.recursivePow(5, 3));
		} catch (AssertionError e) {
			System.out.println("El resultado debería ser 125");
			throw e;
		}
	}

	@Test
	void testIterativePow1() {
		try {
			assertEquals(1, Algoritmos.iterativePow(5, 0));
		} catch (AssertionError e) {
			System.out.println("El resultado debería ser 1");
			throw e;
		}
	}

	@Test
	void testIterativePow2() {
		try {
			assertEquals(125, Algoritmos.iterativePow(5, 3));
		} catch (AssertionError e) {
			System.out.println("El resultado debería ser 125");
			throw e;
		}
	}

	@Test
	void testRecursiveFact1() {
		try {
			assertEquals(1, Algoritmos.recursiveFactorial(0));
		} catch (AssertionError e) {
			System.out.println("El resultado debería ser 1");
			throw e;
		}
	}

	@Test
	void testRecursiveFact2() {
		try {
			assertEquals(120, Algoritmos.recursiveFactorial(5));
		} catch (AssertionError e) {
			System.out.println("El resultado debería ser 120");
			throw e;
		}
	}

	@Test
	void testIterativeFact1() {
		try {
			assertEquals(1, Algoritmos.iterativeFactorial(0));
		} catch (AssertionError e) {
			System.out.println("El resultado debería ser 1");
			throw e;
		}
	}

	@Test
	void testIterativeFact2() {
		try {
			assertEquals(120, Algoritmos.iterativeFactorial(5));
		} catch (AssertionError e) {
			System.out.println("El resultado debería ser 120");
			throw e;
		}
	}
	
	@Test
	void testRecursiveFibo1() {
		try {
			assertEquals(0, Algoritmos.recursiveFibonacci(1));
		} catch (AssertionError e) {
			System.out.println("El resultado debería ser 0");
			throw e;
		}
	}

	@Test
	void testRecursiveFibo2() {
		try {
			assertEquals(13, Algoritmos.recursiveFibonacci(8));
		} catch (AssertionError e) {
			System.out.println("El resultado debería ser 13");
			throw e;
		}
	}
	
	@Test
	void testRecursiveFibo3() {
		try {
			assertEquals(1, Algoritmos.recursiveFibonacci(2));
		} catch (AssertionError e) {
			System.out.println("El resultado debería ser 1");
			throw e;
		}
	}
	
	@Test
	void testIterativeFibo1() {
		try {
			assertEquals(0, Algoritmos.iterativeFibonacci(1));
		} catch (AssertionError e) {
			System.out.println("El resultado debería ser 0");
			throw e;
		}
	}

	@Test
	void testIterativeFibo2() {
		try {
			assertEquals(13, Algoritmos.iterativeFibonacci(8));
		} catch (AssertionError e) {
			System.out.println("El resultado debería ser 13");
			throw e;
		}
	}
	
	@Test
	void testIterativeFibo3() {
		try {
			assertEquals(1, Algoritmos.iterativeFibonacci(2));
		} catch (AssertionError e) {
			System.out.println("El resultado debería ser 1");
			throw e;
		}
	}

}
